+++ CompanyManagement/ViewModels/UserControls/TimeKeepingViewModel.cs
@@ -6,6 +6,8 @@ using System.Windows.Input;
 using CompanyManagement.Database;
 using CompanyManagement.ViewModels.Dialogs.Interfaces;
 using CompanyManagement.ViewModels.UserControls.Interfaces;
+using CompanyManagement.ViewModels.Dialogs;
+using CompanyManagement.Services;
 
 namespace CompanyManagement.ViewModels.UserControls
 {
@@ -64,7 +66,14 @@ namespace CompanyManagement.ViewModels.UserControls
 
         private void ExecuteDeleteCommand(string id)
         {
-            timeKeepingDao.Delete(id);
+            AlertDialogService dialog = new AlertDialogService(
+             "Xóa bảng chấm công",
+             "Bạn chắc chắn muốn xóa bảng chấm công !",
+             () =>
+             {
+                 timeKeepingDao.Delete(id);
+             }, () => { });
+            dialog.Show();
             LoadTimeKeeping();
         }
